+++ ring-core/test/ring/middleware/session/test/cookie.clj
@@ -70,3 +70,10 @@
     (is (not (nil? sess-key)))
     (is (= (read-session store sess-key)
            {:foo now}))))
+
+(deftest cookie-session-deprecation-warning
+  (is (= (with-out-str
+           (binding [*err* *out*]
+             (cookie-store {:key "0123456789abcdef"})))
+         (str "WARNING: The secret key for the session cookie store should be a"
+              " byte array.\nString secret keys have been deprecated.\n"))))