+++ ring-core/test/ring/util/test/request.clj
@@ -43,6 +43,9 @@
     (is (= (character-encoding {:headers {"content-type" "text/plain; charset=UTF-8"}})
            "UTF-8"))
     (is (= (character-encoding {:headers {"content-type" "text/plain;charset=UTF-8"}})
+           "UTF-8")))
+  (testing "content-type with quoted charset"
+    (is (= (character-encoding {:headers {"content-type" "text/plain; charset=\"UTF-8\""}})
            "UTF-8"))))
 
 (deftest test-urlencoded-form?